fix: payment method on timeline
